Moveout TestBootstrapper from ApiApprovals class
